+++ apps/meteor/client/views/admin/subscription/components/cards/SeatsCard.tsx
@@ -1,9 +1,11 @@
+import { Palette } from '@rocket.chat/fuselage';
 import type { ReactElement } from 'react';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 import type { CardProps } from '../FeatureUsageCard';
-import PieGraphCard from '../PieGraphCard';
+import FeatureUsageCard from '../FeatureUsageCard';
+import UsagePieGraph from '../UsagePieGraph';
 
 type SeatsCardProps = {
 	value: number;
@@ -23,11 +25,20 @@ const SeatsCard = ({ value, max }: SeatsCardProps): ReactElement => {
 	const card: CardProps = {
 		title: t('Seats'),
 		infoText: t('Seats_InfoText'),
-		showUpgradeButton: nearLimit,
-		upgradeButtonText: 'Buy_more',
+		...(nearLimit && {
+			upgradeButtonText: t('Buy_more'),
+		}),
 	};
 
-	return <PieGraphCard pieGraph={pieGraph} card={card} />;
+	const seatsLeft = pieGraph.total - pieGraph.used;
+	const color = pieGraph.used / pieGraph.total >= 0.8 ? Palette.statusColor['status-font-on-danger'].toString() : undefined;
+
+	const message = seatsLeft > 0 ? t('Seats_Available', { seatsLeft }) : t('Seats_Required', { seatsRequired: -seatsLeft });
+	return (
+		<FeatureUsageCard card={card}>
+			<UsagePieGraph label={message} used={pieGraph.used} total={pieGraph.total} color={color} />
+		</FeatureUsageCard>
+	);
 };
 
 export default SeatsCard;